requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
bcrypt==4.1.2
PyJWT==2.8.0
python-multipart==0.0.6
